Hadoop 0.18 support along with core API changes.
Renamed c.t.Fields.KEYS to c.t.Fields.GROUP to be consistent with the Cascading model.

Fixed bug where c.t.Tap may inappropriately delete a sink from a task.

Changed c.o.Aggregator to no longer use a Map for the context. Users can now specify custom types by returning
either a new instance from start() or recycling an instance passed into start(). This change will break all existing
implementations of Aggregator. Note, simply returning a new Map<Object,Object> in start() should be sufficient.

Changed all c.o.Function, c.o.Filter, c.o.Aggregator, c.o.ValueAssertion, and c.o.GroupAssertions to accept
a c.f.FlowSession object on all relevant methods. FlowSession provides call-backs into the underyling system
to get configuration properties, fire a "keep alive" ping, or increment a custom counter. This change will
break all existing implemenations of the above interfaces.

Added ability to set serialization tokens via the cascading.serialization.tokens property. This compliments the
c.t.h.SerializationToken annotation.

Optimized co-grouping operation by using c.t.IndexTuple instead of a nested c.t.Tuple.

Changed c.t.Tap and c.s.Scheme sink methods to take a c.t.TupleEntry, instead of c.t.Fields and c.t.Tuple
individually.

Added the c.t.h.SerializationToken Java Annotation. This allows for an int value to be written during serialization
instead of a Class name for custom objects nested in c.t.Tuple instances. This feature should dramatically reduce
the size of Tuples saved in SequenceFiles, and improve the general performance during 'shuffling' between Map and
Reduce stages.

Added c.t.h.TupleSerialization, a Hadoop Serialization implementation. Tuple is no longer Hadoop Writable
and now relies on TupleSerialization for serialization support. Subequently nested objects in c.t.Tuple
only need to be c.l.Comparable. So they can be serialized properly, a Serialization implementation must be
registered with Hadoop. Note all primitive types are handled directly by Tuple, but custom types must
have a Serialization implementation, or must be Hadoop WritableComparable so that the default WritableSerialization
implementation will write them out.

Added support for Hadoop 0.18.x

+++ src/java/cascading/tap/Hfs.java
@@ -29,6 +29,7 @@ import java.util.Map;
 import cascading.scheme.Scheme;
 import cascading.scheme.SequenceFile;
 import cascading.tuple.Fields;
+import cascading.tuple.hadoop.TupleSerialization;
 import cascading.util.Util;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
@@ -270,12 +271,15 @@ public class Hfs extends Tap
     super.sourceInit( conf );
 
     makeLocal( conf, qualifiedPath, "forcing job to local mode, via source: " );
+
+    TupleSerialization.setSerializations( conf ); // allows Hfs to be used independent of Flow
     }
 
   @Override
   public void sinkInit( JobConf conf ) throws IOException
     {
-    if( deleteOnSinkInit )
+    // do not delete if initialized from within a task
+    if( deleteOnSinkInit && conf.get( "mapred.task.partition" ) == null )
       deletePath( conf );
 
     Path qualifiedPath = getQualifiedPath( conf );
@@ -284,6 +288,8 @@ public class Hfs extends Tap
     super.sinkInit( conf );
 
     makeLocal( conf, qualifiedPath, "forcing job to local mode, via sink: " );
+
+    TupleSerialization.setSerializations( conf ); // allows Hfs to be used independent of Flow
     }
 
   private void makeLocal( JobConf conf, Path qualifiedPath, String infoMessage )
@@ -312,7 +318,7 @@ public class Hfs extends Tap
     if( LOG.isDebugEnabled() )
       LOG.debug( "deleting: " + getQualifiedPath( conf ) );
 
-    return getFileSystem( conf ).delete( getPath() );
+    return getFileSystem( conf ).delete( getPath(), true );
     }
 
   @Override
